+++ trunk/variation/src/main/java/org/dishevelled/variation/Variation.java
@@ -61,10 +61,10 @@ public final class Variation
 
     // todo: confirm 1-based; always on 1/+/positive/forward strand
     /** Variation start, using base-counted, one-start (a.k.a. one-based, fully-closed) coordinate system. */
-    private final int start;
+    private final long start;
 
     /** Variation end, using base-counted, one-start (a.k.a. one-based, fully-closed) coordinate system. */
-    private final int end;
+    private final long end;
 
     /** Cached hash code. */
     private final int hashCode;
@@ -88,8 +88,8 @@ public final class Variation
                      final String referenceAllele,
                      final List<String> alternateAlleles,
                      final String region,
-                     final int start,
-                     final int end)
+                     final long start,
+                     final long end)
     {
         checkNotNull(species);
         checkNotNull(reference);
@@ -176,7 +176,7 @@ public final class Variation
      *
      * @return the start for this variation
      */
-    public int getStart()
+    public long getStart()
     {
         return start;
     }
@@ -186,7 +186,7 @@ public final class Variation
      *
      * @return the end for this variation
      */
-    public int getEnd()
+    public long getEnd()
     {
         return end;
     }
